% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Kandinsky_Class_funcs.R
\name{addKanData}
\alias{addKanData}
\alias{addKanData,Seurat-method}
\title{Add Kandinsky data from .rds file to Seurat object}
\usage{
addKanData(object, kanfile = NULL)

\S4method{addKanData}{Seurat}(object, kanfile = NULL)
}
\arguments{
\item{object}{a Seurat object}

\item{kanfile}{character string specifying the path to the file where Kandinsky object is stored. It must be in .rds format}
}
\description{
read .rds file containing Kandinsky object, and add the obejct to already existing Seurat object
}
