% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{AddTxPath}
\alias{AddTxPath}
\title{Add transcript file path to Kandinsky data in a Seurat object}
\usage{
AddTxPath(seurat, path)
}
\arguments{
\item{seurat}{a Seurat object containing Kandinsky data (\code{KanData()})}

\item{path}{character string indicating the path for the transcript coordinates file from CoxMx/Xenium/Merscope platforms}
}
\value{
Seurat object with new \code{tx} slot added to Kandinsky data
}
\description{
Add the full path of CosMx/Xenium/Merscope transcript csv data to Kandinsky data
}
