% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{AlignXeniumCoords}
\alias{AlignXeniumCoords}
\title{align Xenium image and cell coordinates}
\usage{
AlignXeniumCoords(img = NULL, poly = NULL)
}
\arguments{
\item{img}{Xenium image as a multi-layer SpatRaster object}

\item{poly}{Xenium cell polygon data.frame}
}
\value{
list object containing aligned image and cell polygons objects
}
\description{
Adjust Xenium image coordinates to align with cell polygon coordinates
}
