% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Visualization.R
\docType{data}
\name{c12}
\alias{c12}
\title{C12 discrete color palette}
\format{
a vector of 12 colour identifiers
}
\usage{
c12
}
\description{
12-color palette to apply to categorical data
}
\keyword{datasets}
