% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Visualization.R
\docType{data}
\name{c24}
\alias{c24}
\title{C24 discrete color palette}
\format{
a vector of 24 colour identifiers
}
\usage{
c24
}
\description{
24-color palette to apply to categorical data
}
\keyword{datasets}
