% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Colocalisation.R
\name{cellcharter_coloc}
\alias{cellcharter_coloc}
\title{CellCharter colocalisation analysis}
\usage{
cellcharter_coloc(
  data = NULL,
  label = NULL,
  seed = 347548,
  log_fold_change = F,
  python_path = NULL,
  max.cap = 0.9,
  return.mat = T
)
}
\arguments{
\item{data}{a Seurat object containing Kandinsky data slot}

\item{label}{name of the metadata column variable containing cell type annotation to be used for join count test}

\item{seed}{numeric, random seed}

\item{log_fold_change}{boolean, whether to return default CellCharter colocalisation scores (observed - expected) or their log-fold change version (log2(observed/expected))}

\item{python_path}{path of python environment with MESA and its dependencies installed}

\item{max.cap}{numeric, quantile cap to apply to colocalisation coefficients for colour scale creation. Default is 0.9, will set the coefficient value limit to the 90\% quantile of all coefficients}

\item{return.mat}{boolean, whether returning only colocalization plot (FALSE) or also a dataframe containing colocalization test results}
}
\value{
a colocalization plot (when return.mat is FALSE) or a list containing colocalization plot and results in a tabular format
}
\description{
CellCharter is called thrugh reticulate and interacts with Seurat/Kandinsky object after anndata conversion with anndataR
}
\details{
CellCharter must be installed within a python environment together with its dependencies (squidpy)
}
