% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Neighbour_methods.R
\name{centroid_nb}
\alias{centroid_nb}
\title{centroid distance neighbour identification}
\usage{
centroid_nb(poly, d.max = 40)
}
\arguments{
\item{poly}{a sf data frame with polygon geometry}

\item{d.max}{numeric, maximum distance accepted between polygon centroids to define neighbour relation}
}
\value{
weighted neighbour matrix in a listw object from spdep package
}
\description{
Define neighbour relationships according to radial distance threshold between polygon centroids.
}
\details{
Based on dnearneigh function from spdep package
}
\seealso{
Other nb_funcs: 
\code{\link{knn_nb}()},
\code{\link{membrane_nb}()},
\code{\link{nb_expand}()},
\code{\link{nb_update}()},
\code{\link{queen_nb}()}
}
\concept{nb_funcs}
