% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{CosmxFovPlot}
\alias{CosmxFovPlot}
\title{Visualize CosMx IFs staining for a list of FOVs}
\usage{
CosmxFovPlot(fovs = NULL, which = NULL, img_dir = NULL, maxres = 2000)
}
\arguments{
\item{fovs}{data frame containing FOV geometries (square) created via the read_fovfile() function from Kandinsky}

\item{which}{vector of FOV identifier to be considered for plotting}

\item{img_dir}{character stringr specifying the full path of the directory containing Cosmx FOV images (usually named "CellComposite", but other type of images can be selected, such as 'CellOverlay','CellLabels', or 'Morphology2D')}

\item{maxres}{image resolution threshold. Will be applied to each image separately. When several images are selected for plotting, it is advisable to set a low maxres value (within the 500-800 range) to speed up the processing time}
}
\value{
plot of concatenated FOV images
}
\description{
The function return a plot visualizing the immunofluoresence staining images associated with a list of FOVs provided by the user.
When no list is provided, all FOVs represented in the input file will be plotted.
}
\details{
This function is designed to work with CosMx data only.
}
