% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{create_cell_masks}
\alias{create_cell_masks}
\title{Create cell mask polygon}
\usage{
create_cell_masks(
  seurat,
  sample_key = NULL,
  label = NULL,
  label_class = NULL,
  eps = 30,
  minPts = 6,
  concavity = 3,
  lengthThreshold = 0,
  buffer = NULL
)
}
\arguments{
\item{seurat}{a Seurat object containing Kandinsky data}

\item{sample_key}{character string specifying a variable stored in the Seurat object to use as sample/batch annotation.}

\item{label}{character string indicating meta data variable containing cell type annotation}

\item{label_class}{character string indicating which cell type(s) to consider to create the cell mask}

\item{eps}{size (radius) of the epsilon neighborhood. See dbscan documentation for more details.}

\item{minPts}{number of minimum points required in the eps neighborhood for core points (including the point itself). See dbscan documentation for more details.}

\item{concavity}{a relative measure of concavity. 1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1, but they can produce pretty crazy shapes. See concaveman documentation for more details.}

\item{lengthThreshold}{when a segment length is under this threshold, it stops being considered for further detalization. Higher values result in simpler shapes. See concaveman documentation for more details.}

\item{buffer}{numeric, size of spatial buffer to apply to the final cell mask. If NULL, no buffer will be applied. Default is NULL}
}
\value{
a sf data.frame object containing the mask polygon coordinates for each cell cluster defined with dbscan
}
\description{
Create a cell mask polygon based on dbscan spatial clusters of cell type(s) of interest
}
\details{
dbscan spatial density clustering is first applied to define independent clusters of cells.
Concaveman algorithm is then applied separately to each cluster to draw a concave hull mask
}
