% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{gene_contamination_score}
\alias{gene_contamination_score}
\title{Calculate gene-specific contamination score}
\usage{
gene_contamination_score(seurat = NULL, label = NULL, which = NULL)
}
\arguments{
\item{seurat}{Seurat object containing Kandinsky data slot}

\item{label}{character string indicating meta data variable containing cell type annotation}

\item{which}{character vector indicating for which cell type the gene contamination scores will be calculated}
}
\value{
named vector of contamination scores, with each score named with its corresponding gene symbol
}
\description{
For each gene expressed within a dataset, given a cell type of interest,
 this function will calculate a contamination score expressed as the fold change between the average expression of each cell belonging to the defined cell type
 and the average expression of their neighbouring cells defined by Kandinsky belonging to different cell types. A high or low contamination score indicates a lower or higher expression of a gene within a cell than its neighbouring cells, respectively.
}
