% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NB_based_cell_comparison.R
\name{giotto_clust}
\alias{giotto_clust}
\title{Giotto clustering}
\usage{
giotto_clust(
  data = NULL,
  n_clust = 5,
  ngenes = NULL,
  betas = c(0, 5),
  seed = 347548,
  cores = 1,
  python_path = NULL
)
}
\arguments{
\item{data}{Seurat object containing Kandinsky data}

\item{n_clust}{numeric, number of expected clusters. Can be a single number or a numeric vector if multiple cluster numbers needs to be tested}

\item{ngenes}{numeric, number of top variable genes to consider for HMRF domain identification Default is 250. If dataset contains less than 250 genes, all genes will be considered.}

\item{betas}{numeric vector of beta values to use for Giotto HMRF domain identification. Currently only supports two beta values at a time}

\item{seed}{numeric, random seed.}

\item{cores}{numeric, number of cores to use for parallelization}

\item{python_path}{path of python environment with MESA and its dependencies installed}
}
\value{
updated Seurat object with 'giotto_clusters' annotation added to metadata slot for each beta value
}
\description{
Giotto clustering performed by calling the R package Giotto through Kandinsky
}
\details{
Giotto must be installed within the same environmnent of Kandinsky
}
