% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{grid_visium}
\alias{grid_visium}
\title{Create Visium-like spots in a ordered grid}
\usage{
grid_visium(data = NULL)
}
\arguments{
\item{data}{a Seurat object containing Kandinsky data slot}
}
\value{
VisiumHD-like circular spot polygon
}
\description{
Create Visium-like circular spots covering the whole sample area
}
