% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{grid_visiumhd}
\alias{grid_visiumhd}
\title{Create VisiumHD-like bins in a ordered grid}
\usage{
grid_visiumhd(data = NULL, res = NULL)
}
\arguments{
\item{data}{a Seurat object containing Kandinsky data slot}

\item{res}{numeric, Visium-HD bin resolution when \code{style} is set to 'bins'.}
}
\value{
VisiumHD-like square bin polygons
}
\description{
Split sample area into a regular square grid to replicate VisiumHD bins
}
