% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NB_based_cell_comparison.R
\name{hoodscanR_clust}
\alias{hoodscanR_clust}
\title{hoodscanR clustering}
\usage{
hoodscanR_clust(
  data = NULL,
  k = 100,
  label = NULL,
  n_clust = 5,
  sample_key = NULL
)
}
\arguments{
\item{data}{Seurat object containing Kandinsky data}

\item{k}{numeric, number of nearest neighbours to consider for hoodscanR clustering}

\item{label}{character string specifying the variable name to be used to defne cell annotation groups}

\item{n_clust}{numeric, number of expected clusters. Can be a single number or a numeric vector if multiple cluster numbers needs to be tested}

\item{sample_key}{character string specifying a variable stored in the Seurat object to use as sample/batch annotation. If not NULL, neighbour networks will be defined separately for each sample/batch. Default is NULL}
}
\value{
updated Seurat object with 'hoodscanR_clusters' annotation added to metadata slot
}
\description{
hoodscanR clustering performed by calling the R package hoodscanR through Kandinsky
}
\details{
hoodscanR must be installed within the same environmnent of Kandinsky
}
