% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Colocalisation.R
\name{hoodscanR_coloc}
\alias{hoodscanR_coloc}
\title{hoodscanR colocalisation}
\usage{
hoodscanR_coloc(
  data = NULL,
  k = 100,
  label = NULL,
  sample_key = NULL,
  max.cap = 1,
  return.mat = T
)
}
\arguments{
\item{data}{Seurat object containing Kandinsky data}

\item{k}{numeric, number of nearest neighbours to consider for hoodscanR clustering}

\item{label}{character string specifying the variable name to be used to defne cell annotation groups}

\item{sample_key}{character string specifying a variable stored in the Seurat object to use as sample/batch annotation. If not NULL, neighbour networks will be defined separately for each sample/batch. Default is NULL}

\item{max.cap}{numeric, quantile cap to apply to colocalisation coefficients for colour scale creation. Default is 1, will set the coefficient value limit to the highest value of all coefficients}

\item{return.mat}{boolean, whether returning only colocalization plot (FALSE) or also a dataframe containing colocalization test results}
}
\value{
a colocalization plot (when return.mat is FALSE) or a list containing colocalization plot and results in a tabular format
}
\description{
hoodscanR colocalisation performed by calling the R package hoodscanR through Kandinsky
}
\details{
hoodscanR must be installed within the same environmnent of Kandinsky..
}
