% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Hot_cold_spot_analysis_.R
\name{hotspot_count}
\alias{hotspot_count}
\title{Count and describe Getis-Ord Gi* hotspot sites}
\usage{
hotspot_count(data = NULL, feature = NULL, ...)
}
\arguments{
\item{data}{a Seurat object containing Kandinsky data (\code{KanData()})}

\item{feature}{character string specifying the name of the feature(s) used to compute Getis-Ord Gi* statistics}

\item{...}{arguments passed to hotspot_analysis function}
}
\value{
data.frame summarising hotspot and coldspot size and number for each variable of interest
}
\description{
Summarise number and min/max/median size of hot- and cold-spots for a variable of choice.
}
\details{
This function will look for the presence of already existing hotspot annotation in the meta.data of Seurat object for the variable(s) of interest.
If no annotations exist, the \code{hotspot_analysis()} function will be automatically called.
}
