% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Colocalisation.R
\name{jc_coloc}
\alias{jc_coloc}
\title{JC colocalization test and plot using Seurat/Kandinsky object}
\usage{
jc_coloc(
  data = NULL,
  label = NULL,
  max.cap = 0.9,
  return.mat = F,
  wc.type = c("mat", "listw")
)
}
\arguments{
\item{data}{a Seurat object containing Kandinsky data slot}

\item{label}{name of the metadata column variable containing cell type annotation to be used for join count test}

\item{max.cap}{numeric, quantile cap to apply to colocalisation coefficients for colour scale creation. Default is 0.9, will set the coefficient value limit to the 90\% quantile of all coefficients}

\item{return.mat}{boolean, whether returning only colocalization plot (FALSE) or also a dataframe containing colocalization test results}

\item{wc.type}{input type to use to extract summary weight constants used to calculate join count statistics. Must be one of the following: "listw", "mat"}
}
\value{
a colocalization plot (when return.mat is FALSE) or a list containing colocalization plot and results in a tabular format
}
\description{
Call multi_jc() function to perform colocalization test, and plot results in the form of a heatmap
}
\seealso{
\code{\link[=squidpy_coloc]{squidpy_coloc()}}, \code{\link[=cellcharter_coloc]{cellcharter_coloc()}}, \code{\link[=hoodscanR_coloc]{hoodscanR_coloc()}}, \code{\link[=giotto_coloc]{giotto_coloc()}} for alternative colocalisation methods
}
