% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Kandinsky_Class_funcs.R
\docType{class}
\name{Kandinsky-class}
\alias{Kandinsky-class}
\alias{Kandinsky}
\title{Kandinsky S4 object class}
\description{
Kandinsky is a list object storing spatial data and information for spatial transcriptomic dataset analysed through Seurat.
}
\section{Slots}{

\describe{
\item{\code{platform}}{sequencing technology}

\item{\code{nb}}{weighted neighbour list object created with the package\code{spdep}}

\item{\code{sf}}{geometry data frame created with the package \code{sf} and containing cell/spot coordinates}

\item{\code{nb.type}}{character indicating the neighbouring strategy used to initialize Kandinsky object. Must be one of the following: \code{Q}, \code{K}, \code{C}, \code{M}}

\item{\code{nnMat}}{list object containing nearest neighbour matrix created with one of the \code{nnMat} functions implemented in Kandinsky}

\item{\code{spot_distance}}{numeric value indicating centroid-to-centroid distance between Visium spots or Visium HD bins}

\item{\code{texture}}{data.frame containing GLCM texture features extracted from Visium H&E slide image}

\item{\code{tx}}{character string indicating the path for the transcript coordinates file from CoxMx/Xenium/Merscope platforms}

\item{\code{fov_mask}}{geometry data frame created with the package \code{sf} and containing CosMx FOVs bounding box coordinates}

\item{\code{img}}{Visium H&E or any other tissue slide image stored as a SpatRaster or SpatrasterCollection object via the package \code{terra}}

\item{\code{mask}}{masked H&E image obtained via the Kandinsky function \code{he_mask} and stored as a SpatRaster or SpatrasterCollection object via the package \code{terra}}
}}

