% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kandinsky.R
\name{kandinsky}
\alias{kandinsky}
\title{Turn any dataset into a Kandinsky painting}
\usage{
kandinsky(df = NULL, rv = runif(1000))
}
\arguments{
\item{df}{a dataset (tested: \code{ts} object, \code{data.frame}, \code{tibble})}

\item{rv}{a numeric vector in the range 0 to 1, if \code{df} is not NULL,
default is \code{runif(1000)}}
}
\description{
A function to turn any numeric vector in the range 0 to 1, or a given dataset,
into a Kandinsky-like painting, with a few common Kandinsky shapes, e.g.
rectangles and circles, drawn with wrappers around the \code{grid} package.
If the input is a dataset, the function will first convert it into a numeric
vector in the range 0 to 1.
}
\details{
In order to convert a dataset (\code{ts} object, \code{data.frame}, \code{tibble})
into a numeric vector in the range 0 to 1, the function first converts the dataset
character or factor columns into numeric columns, then normalizes all (now
numeric) columns to be in the 0 to 1 range, then vectorizes the dataset into
a single vector by row.
}
\note{
This function is currently tested on a Windows 7 machine only
}
\examples{
kandinsky(airquality)

}
\seealso{
\link{http://giorasimchoni.com/}
}
