% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NB_based_cell_comparison.R
\name{kmeans_clust}
\alias{kmeans_clust}
\title{K-Means clustering}
\usage{
kmeans_clust(data = NULL, n_clust = 1:10, label = NULL, seed = 347548)
}
\arguments{
\item{data}{Seurat object containing Kandinsky data}

\item{n_clust}{numeric, number of expected clusters. Can be a single number or a numeric vector if multiple cluster numbers needs to be tested}

\item{label}{character string specifying the variable name to be used to defne cell annotation groups}

\item{seed}{numeric, random seed.}
}
\value{
updated Seurat object with 'kmeans_clusters' annotation added to metadata slot
}
\description{
ISCHIA-inspired K-Means clustering
}
\details{
When more than one cluster configurations are provided, the function chooses the one minimizing Within-Cluster Sum of Squares (WSS), as implemented in the R package ISCHIA
}
