% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Neighbour_methods.R
\name{knn_nb}
\alias{knn_nb}
\title{K-nearest neighbour neighbour identification}
\usage{
knn_nb(poly = NULL, k = 10)
}
\arguments{
\item{poly}{a sf data frame with polygon geometry}

\item{k}{numeric, number of nearest neighbour to select with knn algorithm}
}
\value{
weighted neighbour matrix in a listw object from spdep package
}
\description{
Define neighbour relationships according to K-nearest neighbour algorithm.
}
\details{
based on knearneigh function from spdep package
}
\seealso{
Other nb_funcs: 
\code{\link{centroid_nb}()},
\code{\link{membrane_nb}()},
\code{\link{nb_expand}()},
\code{\link{nb_update}()},
\code{\link{queen_nb}()},
\code{\link{tri_nb}()}
}
\concept{nb_funcs}
