% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{LoadXeniumImage}
\alias{LoadXeniumImage}
\title{Load Xenium OME tiff image}
\usage{
LoadXeniumImage(
  path = "/Volumes/colcc-ITH2_scRNA_datasets/data/10X_Xenium_CRC_public/Breast_5K/",
  channel_cols = c("gray70", "limegreen", "cyan", "indianred3"),
  res = 5
)
}
\arguments{
\item{path}{character string indicating the path of the Xenium OME tiff image}

\item{channel_cols}{vector of color names to be assigned to each of the OME image channel}

\item{res}{integer specifying resolution levels to read.}
}
\value{
Xenium image as a multi-layer SpatRaster object
}
\description{
Load morphology OME tiff images from Xenium output with the help of RBioFormats read.image() function
}
