% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Neighbour_methods.R
\name{membrane_nb}
\alias{membrane_nb}
\title{membrane-based neighbour identification}
\usage{
membrane_nb(poly, d.max = 0)
}
\arguments{
\item{poly}{a sf data frame with polygon geometry}

\item{d.max}{numeric, maximum distance accepted between polygon borders to define neighbour relation}
}
\value{
weighted neighbour matrix in a listw object from spdep package
}
\description{
Define neighbour relationships according to contact distance treshold between polygon borders.
}
\details{
Based on polygon intersection check done with st_intersects() function from sf package.
When minimum distance is higher than 0, a round buffer of size max.dist/2 is applied to all polygons before calling st_intersects()
}
\seealso{
Other nb_funcs: 
\code{\link{centroid_nb}()},
\code{\link{knn_nb}()},
\code{\link{nb_expand}()},
\code{\link{nb_update}()},
\code{\link{queen_nb}()},
\code{\link{tri_nb}()}
}
\concept{nb_funcs}
