% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NB_based_cell_comparison.R
\name{mesa_clust}
\alias{mesa_clust}
\title{Reticulate-based MESA clustering}
\usage{
mesa_clust(
  data = NULL,
  k = 10,
  n_clust = 10,
  seed = 347548,
  python_path = NULL
)
}
\arguments{
\item{data}{a Seurat object containing Kandinsky data}

\item{k}{numeric, number of nearest neighbours to consider for MESA gene expression aggregation}

\item{n_clust}{numeric, number of expected clusters}

\item{seed}{numeric, random seed}

\item{python_path}{path of python environment with MESA and its dependencies installed}
}
\value{
updated Seurat object with 'mesa_clusters' annotation added to metadata slot
}
\description{
MESA is called thrugh reticulate and interacts with Seurat/Kandinsky object after anndata conversion with anndataR
}
\details{
MESA must be installed within a python environment together with its dependencies (scanpy,sklearn.cluster)
}
