% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Colocalisation.R
\name{multi_jc}
\alias{multi_jc}
\title{same-color (BB) and different-color (BW) join count test}
\usage{
multi_jc(fx, listw, wc.type = c("mat", "listw"))
}
\arguments{
\item{fx}{a factor of the same length as the neighbours and weights objects in listw}

\item{listw}{a listw object}

\item{wc.type}{input type to use to extract summary weight constants used to calculate join count statistics. Must be one of the following: "listw", "mat"}
}
\value{
data.frame containing join count coefficients for each cell type pair tested
}
\description{
Readapted version of joincount.multi() function from the spdep package.
}
