% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NB_based_cell_comparison.R
\name{nb_entropy}
\alias{nb_entropy}
\title{Calculate Shannon entropy of cell neighbourhood composition}
\usage{
nb_entropy(seurat)
}
\arguments{
\item{seurat}{seurat object containing Kandinsky data slot with pre-computed neighbourhood composition matrix in \code{nnMat} slot}
}
\value{
seurat object with new metadata variable reporting neighbourhood entropy for each cell
}
\description{
Given cell type composition of single-cell neighbourhood, calculate Shannon entropy associated with each neighbourhood
}
