% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Neighbour_methods.R
\name{nb_expand}
\alias{nb_expand}
\title{expand neighbour links to N higher orders}
\usage{
nb_expand(seurat, maxorder = 2, cumul = T)
}
\arguments{
\item{seurat}{a Seurat object containing Kandinsky data slot}

\item{maxorder}{numeric, link distance considered to define higher order neighbours}

\item{cumul}{boolean, whether returning updated neighbour network with only higher order relationships (FALSE) or with the cumulative union of first and higher order neighbour links (TRUE)}
}
\value{
seurat object with updated 'nb' neighbours in Kandinsky data
}
\description{
Given a Kandinsky object containing a neighbour network, this function will identify, for each cell/node, higher order neighbours that are distant n (maxorder) number of links from each other.
For more information, see nblag() function in the spdep package.
}
\seealso{
Other nb_funcs: 
\code{\link{centroid_nb}()},
\code{\link{knn_nb}()},
\code{\link{membrane_nb}()},
\code{\link{nb_update}()},
\code{\link{queen_nb}()}
}
\concept{nb_funcs}
