% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NB_based_cell_comparison.R
\name{nb_lm}
\alias{nb_lm}
\title{Neighbourhood-based multiple linear regression of gene expression values}
\usage{
nb_lm(
  seurat,
  label = NULL,
  which = NULL,
  features = NULL,
  layer = "data",
  rand = NULL,
  minprop = 0.1
)
}
\arguments{
\item{seurat}{a Seurat object containing Kandinsky data slot}

\item{label}{character string specifying the name of the variable to be used for cell type annotation.}

\item{which}{name of the cell type to consider for fitting linear models. Currently only one cell type is accepted as input}

\item{features}{character vector containing gene identifiers to be fitted}

\item{layer}{name of Seurat layer to use for gene expression data}

\item{rand}{name of random variables to be included in a mixed effect model.}

\item{minprop}{numeric, genes expressed in less than this proportion of cells will be excluded from the analysis}
}
\value{
data frame reporting, for each gene, the regression coefficient and p-value associated with each cell type
}
\description{
Fit a multiple linear regression for a gene of interest using neighbourhood cell count for each cell type as independent variables
}
\details{
This function is based on fastLm() function from RcppEigen pakage. In case user wanted to include one or more confounfing (random) variables into the model,
the function lmer() in instead used to fit a mixed effect model. In both cases, the output is a data frame reporting regression coefficient and pvalue for each independent variable.
}
