% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Neighbour_methods.R
\name{nb_summary}
\alias{nb_summary}
\title{Get Kandinsky neighbourhood summary metrics}
\usage{
nb_summary(seurat = NULL)
}
\arguments{
\item{seurat}{a Seurat object containing Kandinsky data}
}
\value{
text summary of neighbour metrics
}
\description{
Print a summary of neighbour counts and sizes detected by Kandinsky
}
