% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Visualization.R
\name{nbSizePlot}
\alias{nbSizePlot}
\title{density plot of neighbourhood size across cells/spots}
\usage{
nbSizePlot(seurat)
}
\arguments{
\item{seurat}{a Seurat object containing Kandinsky data}
}
\value{
density plot of cell/spot neighbourhood sizes
}
\description{
Visualize the distribution of possible neighbourhood sizes observed across cells/spots within the dataset
}
