% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{populate_sf}
\alias{populate_sf}
\title{Add variables to sf object stored in Kandinsky data}
\usage{
populate_sf(
  seurat = NULL,
  vars = NULL,
  assay = DefaultAssay(seurat),
  layer = "data",
  return.seurat = F
)
}
\arguments{
\item{seurat}{a Seurat object containing Kandinsky data (\code{KanData()})}

\item{vars}{character vector containing variable names stored in Seurat object to be copied into the KanData \code{sf} object}

\item{assay}{name of the Seurat assay to be used to search for variables of interest}

\item{layer}{name of the Seurat layer to be used to search for variables of interest}

\item{return.seurat}{boolean, set to FALSE by default. It indicates whether returning Seurat object with updated KanData \code{sf} containing the new variables (TRUE), or the updated KanData \code{sf} object only}
}
\value{
if return.seurat is set to TRUE, a Seurat object with updated KanData \code{sf} containing the new variables. if return.seurat is set to FALSE, only the updated KanData \code{sf} object will be returned
}
\description{
Add information to the Kandinsky 'sf' slot. Based on Seurat::FetchData function
}
