% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Prepare_seurat_data.R
\name{prepare_cosmx_seurat}
\alias{prepare_cosmx_seurat}
\title{Prepare NanoString Cosmx Seurat data}
\usage{
prepare_cosmx_seurat(
  path = NULL,
  dataset.id = "X",
  pattern = NULL,
  fovs = NULL
)
}
\arguments{
\item{path}{character string specifying the path to the CosMx input files directory}

\item{dataset.id}{character string that will be used to name the output Seurat object identity}

\item{pattern}{character string that will be used as a key to identify CosMx input file names in case when input files from multiple CosMx samples/slides are stored in the same directory}

\item{fovs}{vector of FOV identifiers to subset for downstream analysis.}
}
\value{
a Seurat object containing CosMx count matrix, metadata, and additional files/file paths to be used for downstream analysis with Kandinsky functions
}
\description{
Initial formatting of CosMx data to work with Seurat and Kandinsky
}
\details{
This function will use NanoString CosMx raw input data to build a new Seurat object. All Cosmx input files must be stored in the same folder that will be specified to call the function
}
\seealso{
Other prepare_data: 
\code{\link{prepare_g4x_seurat}()},
\code{\link{prepare_merscope_seurat}()},
\code{\link{prepare_proseg_seurat}()},
\code{\link{prepare_slideseq_seurat}()},
\code{\link{prepare_visiumHD_seurat}()},
\code{\link{prepare_visium_seurat}()},
\code{\link{prepare_xenium_seurat}()}
}
\concept{prepare_data}
