% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Prepare_seurat_data.R
\name{prepare_g4x_seurat}
\alias{prepare_g4x_seurat}
\title{Prepare Singular Genomics G4X Seurat data}
\usage{
prepare_g4x_seurat(
  path = NULL,
  dataset.id = NULL,
  pattern = NULL,
  h5 = T,
  segmentation = F
)
}
\arguments{
\item{path}{character string specifying the path to the G4X input files directory}

\item{dataset.id}{character string that will be used to name the output Seurat object identity}

\item{pattern}{character string that will be used as a key to identify G4X input file names in case when input files from multiple G4X samples/slides are stored in the same directory}

\item{h5}{boolean, whether reading count matrix and metadata from 'feature_matrix.h5' (TRUE) file or remaining files in the 'single_cell_data/' folder (FALSE)}

\item{segmentation}{boolean, whether building cell polygons using G4X segmentation masks from .npz files (TRUE) or using cell centroids (FALSE). Reading .npz files requires the use of python numpy library through reticulate}
}
\value{
a Seurat object containing G4X count matrix, metadata, and additional files/file paths to be used for downstream analysis with Kandinsky functions
}
\description{
Initial formatting of Singular Genomics G4X data to work with Seurat and Kandinsky
}
\details{
This function will use Singular Genomics G4X raw input data to build a new Seurat object. All G4X input files must be stored in the same folder that will be specified to call the function
}
\seealso{
Other prepare_data: 
\code{\link{prepare_cosmx_seurat}()},
\code{\link{prepare_merscope_seurat}()},
\code{\link{prepare_proseg_seurat}()},
\code{\link{prepare_slideseq_seurat}()},
\code{\link{prepare_visiumHD_seurat}()},
\code{\link{prepare_visium_seurat}()},
\code{\link{prepare_xenium_seurat}()}
}
\concept{prepare_data}
