% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Prepare_seurat_data.R
\name{prepare_proseg_seurat}
\alias{prepare_proseg_seurat}
\title{Prepare proseg Seurat data}
\usage{
prepare_proseg_seurat(
  path = NULL,
  dataset.id = NULL,
  pattern = NULL,
  fovs = NULL,
  tech = c("cosmx", "merscope", "xenium")
)
}
\arguments{
\item{path}{character string specifying the path to the proseg input files directory}

\item{dataset.id}{character string that will be used to name the output Seurat object identity}

\item{pattern}{character string that will be used as a key to identify proseg input file names in case when input files from multiple proseg samples/slides are stored in the same directory}

\item{fovs}{vector of FOV identifiers to subset for downstream analysis.}

\item{tech}{character string specifying the technology used to generate the data processed with proseg. Must be one of the following: "cosmx", "merscope", "xenium"}
}
\value{
a Seurat object containing proseg expected count matrix, metadata, and additional files/file paths to be used for downstream analysis with Kandinsky functions
}
\description{
Initial formatting of proseg re-segmented data to work with Seurat and Kandinsky
}
\details{
This function will use proseg raw input data (generated from either CosMx, Xenium, or Merscope datasets) to build a new Seurat object. All proseg input files must be stored in the same folder that will be specified to call the function
}
\seealso{
Other prepare_data: 
\code{\link{prepare_cosmx_seurat}()},
\code{\link{prepare_g4x_seurat}()},
\code{\link{prepare_merscope_seurat}()},
\code{\link{prepare_slideseq_seurat}()},
\code{\link{prepare_visiumHD_seurat}()},
\code{\link{prepare_visium_seurat}()},
\code{\link{prepare_xenium_seurat}()}
}
\concept{prepare_data}
