% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Prepare_seurat_data.R
\name{prepare_xenium_seurat}
\alias{prepare_xenium_seurat}
\title{Prepare 10X Xenium Seurat data}
\usage{
prepare_xenium_seurat(path = NULL, dataset.id = "X", h5 = TRUE)
}
\arguments{
\item{path}{character string specifying the path to the Xenium input files directory}

\item{dataset.id}{character string that will be used to name the output Seurat object identity}

\item{h5}{boolean, whether reading count matrix from 'cell_feature_matrix.h5' (TRUE) file or zipped 'cell_feature_matrix.tar.gz' folder (FALSE)}
}
\value{
a Seurat object containing Xenium count matrix, metadata, and additional files/file paths to be used for downstream analysis with Kandinsky functions
}
\description{
Initial formatting of Xenium data to work with Seurat and Kandinsky
}
\details{
This function will use 10X Xenium raw input data to build a new Seurat object. All Xenium input files must be stored in the same folder that will be specified to call the function
}
\seealso{
Other prepare_data: 
\code{\link{prepare_cosmx_seurat}()},
\code{\link{prepare_g4x_seurat}()},
\code{\link{prepare_merscope_seurat}()},
\code{\link{prepare_proseg_seurat}()},
\code{\link{prepare_slideseq_seurat}()},
\code{\link{prepare_visiumHD_seurat}()},
\code{\link{prepare_visium_seurat}()}
}
\concept{prepare_data}
