% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{read_fovfile}
\alias{read_fovfile}
\title{Read CosMx fov position file}
\usage{
read_fovfile(fov_path = NULL, buffer = 255, fov_ids = NULL)
}
\arguments{
\item{fov_path}{character string indicating the path for the fov coordinates file from CoxMx platform}

\item{buffer}{numeric value (micrometers) indicating the radial expansion that will be applied to FOV centroid to reconstruct the FOV full size polygon.
It should not be modified, as this is the correct parameter to get the standard CosMx FOV size,}

\item{fov_ids}{vector of FOV identifier to be considered for the final polygon output. If set to \code{NULL}, all the FOVs reported in the input file will be used}
}
\value{
a sf data frame containing a geometry column with all FOV polygon coordinates
}
\description{
Read fov position csv file generated by NanoString CosMx platform
}
\details{
The function will create a polygon for each CosMx field of view reported in the input file, with coordinates
measured in micrometers. The function assumed a standardized FOV size of 0.51x0.51 millimeters, and it does not account
for any possible position offset compared to cell polygon coordinates.
The output FOV polygons can be used in principle as a reference to arrange CosMx tif/png images as background for data visualization
}
