% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{resample_cells}
\alias{resample_cells}
\title{resample single cells in Seurat object}
\usage{
resample_cells(
  seurat = NULL,
  label = "cell_types",
  spatial = T,
  maxcells = 10000,
  seed = 347548,
  return.seurat = T,
  update.kandinsky = T
)
}
\arguments{
\item{seurat}{Seurat object}

\item{label}{character string indicating meta data variable containing cell type annotation}

\item{spatial}{boolean, whether or not applying a spatially-informed resampling}

\item{maxcells}{numeric, ideal number of cells to resample.
When \code{spatial} is true, an optimal number of cells as close as possible to \code{maxcells} argument
will be sampled by testing multiple resampling resolution parameters}

\item{seed}{numeric, random seed for reproducibility}

\item{return.seurat}{boolean, whether returning resampled IDs (FALSE) or the whole seurat object already filtered with resampled IDs (TRUE)}

\item{update.kandinsky}{boolean, whether or not update Kandinsky data after resampling. Default is TRUE. Only applied when \code{return.seurat} is TRUE}
}
\value{
vector of resampled cell IDs or Seurat object with resampled cell subset
}
\description{
Apply a random resample of cell identifier in Seurat object. Original cell type proportion in Seurat dataset will be preserved
when non-spatial sampling strategy is applied. With spatial resampling, rare/sparse cell types will be less penalised from resampling
compared to more abundant and spatially dense cell type, with a sampling proportion that tends to be
inversely correlated with the initial cell type relative abundance
}
