% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Visualization.R
\name{SlidePlot}
\alias{SlidePlot}
\title{Plot Kandinsky images}
\usage{
SlidePlot(
  seurat = NULL,
  max.px = 10^6,
  img_type = c("img", "mask"),
  as.ggplot = T
)
}
\arguments{
\item{seurat}{a Seurat object containing Kandinsky data (\code{KanData()})}

\item{max.px}{numeric value indicating the number of image pixels to plot. Higher values will give a better resolution and a  heavier image size}

\item{img_type}{type of raster image stored in Kandinsky data. Must be one of the following: \code{img}, \code{mask}}

\item{as.ggplot, }{boolean, whether returning a plot created with \code{tidyterra::geom_spatraster_rgb()} function (TRUE) or using \code{terra::plotRGB()} function}
}
\value{
Kandinsky image plot, as a base R plot (as.ggplot=F) or as a ggplot object (as.ggplot=T)
}
\description{
Visualize raster images stored in Kandinsky data
}
