% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NB_based_cell_comparison.R
\name{spatialleiden_clust}
\alias{spatialleiden_clust}
\title{Reticulate-based SpatialLeiden clustering}
\usage{
spatialleiden_clust(
  data = NULL,
  n_clust = NULL,
  nfeatures = 2000,
  seed = 347548,
  python_path = NULL,
  resolution = 1,
  embedding = NULL,
  layer_ratio = 1.5
)
}
\arguments{
\item{data}{a Seurat object containing Kandinsky data}

\item{n_clust}{numeric, number of expected clusters}

\item{nfeatures}{numeric, number of variable features to use for dimensionality reduction (PCA)}

\item{seed}{numeric, random seed}

\item{python_path}{path of python environment with SpatialLeiden and its dependencies installed}

\item{resolution}{numeric, resolution for the latent space and spatial layer, respectively.}

\item{embedding}{name of dimension reduction embedding to use for SpatialLeiden clustering. If NULL, the function will run PCA with scanpy through reticulate. Default is NULL.}

\item{layer_ratio}{numeric, the ratio of the weighting of the layers; latent space vs spatial. A higher ratio will increase relevance of the spatial neighbors and lead to more spatially homogeneous clusters}
}
\value{
updated Seurat object with 'spatialleiden_clusters' annotation added to metadata slot
}
\description{
SpatialLeiden is called thrugh reticulate and interacts with Seurat/Kandinsky object after anndata conversion with anndataR
}
\details{
SpatialLeiden must be installed within a python environment together with its dependencies (squidpy,scanpy,leidenalg)
}
