% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NB_based_cell_comparison.R
\name{spatopic_clust}
\alias{spatopic_clust}
\title{SpaTopic clustering}
\usage{
spatopic_clust(
  data = NULL,
  label = NULL,
  sample_key = NULL,
  n_clust = 2:9,
  sigma = 25,
  region_radius = 500,
  kneigh = 5,
  trace = TRUE,
  thin = 20,
  burnin = 1000,
  niter = 200,
  seed = 347548
)
}
\arguments{
\item{data}{Seurat object containing Kandinsky data}

\item{label}{character string specifying the variable name to be used to defne cell annotation groups}

\item{sample_key}{character string specifying a variable stored in the Seurat object to use as sample/batch annotation. If not NULL, neighbour networks will be defined separately for each sample/batch. Default is NULL.}

\item{n_clust}{numeric, number of expected clusters. Can be a single number or a numeric vector if multiple cluster numbers needs to be tested}

\item{sigma}{Default is 50. The lengthscale of the Nearest-neighbor Exponential Kernel. Sigma controls the strength of decay of correlation with distance in the kernel function. Please check the paper for more information. Need to be adjusted based on the image resolution}

\item{region_radius}{Default is 400. The radius for each grid square when sampling region centers for each image. Need to be adjusted based on the image resolution and pattern complexity.}

\item{kneigh}{Default is 5. Only consider the top 5 closest region centers for each cell.}

\item{trace}{Default is TRUE Compute and save log likelihood, Ndk, Nwk for every posterior samples. Useful when you want to use DIC to select number of topics, but it is time consuming to compute the likelihood for every posterior samples.}

\item{thin}{Default is 20. Key parameter in Gibbs sampling. Collect a posterior sample for every thin=20 iterations.}

\item{burnin}{Default is 1000. Key parameter in Gibbs sampling. Start to collect posterior samples after 1000 iterations. You may increase the number of iterations for burn-in for highly complex tissue images.}

\item{niter}{Default is 200. Key parameter in Gibbs sampling. Number of posterior samples collected for model inference.}

\item{seed}{Default is 347548 Random seed.}
}
\value{
updated Seurat object with 'spatopic_clusters' annotation added to metadata slot
}
\description{
SpatialTopic clustering performed by calling the R package SpaTopic through Kandinsky
}
\details{
SpaTopic must be installed within the same environmnent of Kandinsky
}
