% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Colocalisation.R
\name{spweights_mat}
\alias{spweights_mat}
\title{estimate weight summary constansts from spatial weights matrix}
\usage{
spweights_mat(listw)
}
\arguments{
\item{listw}{a weighted network in list (listw) or matrix (CsparseMatrix) format}
}
\value{
a list obejct containing summary constants
}
\description{
Faster extraction of neighbour network summary weights (taken from spweights.constants function in spdep )
}
\details{
For more information see spweights.constants() function documentation
}
