% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{stitch_samples}
\alias{stitch_samples}
\title{Merge coordinate reference systems from independent samples}
\usage{
stitch_samples(
  data = NULL,
  x = NULL,
  y = NULL,
  sample.id = NULL,
  max_rows = NULL,
  max_cols = NULL
)
}
\arguments{
\item{data}{a data frame containing x and y coordinates into separate columns}

\item{x}{name of the x coordinate column}

\item{y}{name of the y coordinate column}

\item{sample.id}{character string specifying the name of the column storing sample identifiers}

\item{max_rows}{numeric, maximum number of samples to be disposed on a same column of the final sample grid. If set to \code{NULL}, maximum number of rows will be automatically calculated to obtain a square grid}

\item{max_cols}{numeric, maximum number of samples to be disposed on a same row of the final sample grid. If set to \code{NULL}, maximum number of columns will be automatically calculated to obtain a square grid}
}
\value{
data frame object with new stitched 'x_global' and 'y_global' coordinate columns
}
\description{
Combine samples coming from independent experiments by stitching cells/spots coordinates through x/y coordinate shifting into a single grid
}
