% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{subsample_nb}
\alias{subsample_nb}
\title{subsample neighbour network links}
\usage{
subsample_nb(seurat = NULL, exp_links = 6, seed = 347548)
}
\arguments{
\item{seurat}{a Seurat object containing Kandinsky data slot}

\item{exp_links}{numeric, maximum number of expected neighbours per cell after the subsetting}

\item{seed}{numeric, random seed for reproducibility}
}
\value{
updated seurat object network in a listw format
}
\description{
Given a Kandinsky neighbour network in a listw format, this function randomly subset
the number of neighbours assigned to each cell up to a user-defined maximum number of neighbours.
}
\details{
In order to be compatible with most of Kandinsky functions (due to its dependency on spdep package), the resulting
neighbour matrix must be symmetric even after the subsetting. Therefore, an additional step to ensure matrix symmetry
is applied after the subsetting, and for this reason some cells might still have a number of neighbours higher than the user-provided threshold.
}
