% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Neighbour_methods.R
\name{tri_nb}
\alias{tri_nb}
\title{Delaunay triangulation neighbour identification}
\usage{
tri_nb(poly, soi = F)
}
\arguments{
\item{poly}{a sf data frame with polygon geometry}

\item{soi}{boolean, whether or not filter Delaunay network to keep sphere of influence (SOI) graph. Default is FALSE}
}
\value{
weighted neighbour matrix in a listw object from spdep package
}
\description{
Define neighbour relationships according to Delaunay triangulation network
}
\details{
based on tri.mesh function from tripack package.
}
\seealso{
Other nb_funcs: 
\code{\link{centroid_nb}()},
\code{\link{knn_nb}()},
\code{\link{membrane_nb}()},
\code{\link{nb_expand}()},
\code{\link{nb_update}()},
\code{\link{queen_nb}()}
}
\concept{nb_funcs}
