% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Kandinsky_Class_funcs.R
\name{update_kandinsky}
\alias{update_kandinsky}
\alias{update_kandinsky,Seurat-method}
\alias{update_kandinsky,}
\alias{Seurat_update_kandinsky}
\title{Update Kandinsky data}
\usage{
update_kandinsky(seurat = NULL)

\S4method{update_kandinsky}{Seurat}(seurat = NULL)
}
\arguments{
\item{seurat}{a Seurat object containing Kandinsky data (\code{KanData()})}
}
\value{
Seurat object with updated Kandinsky data
}
\description{
data objects included in the KanData slots based on the real Seurat object.
By default Kandinsky data will not be subsetted together with Seurat during normal filtering steps using \code{subset()} or row/column names subset (e.g. \code{seurat[,ids]}).
Use \code{update_kandinsky()} to check and fix discrepancies between Seurat and KanData cell/spot ids
}
