% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{visiumize}
\alias{visiumize}
\title{Convert single-cell spatial transcriptomics data to Visium/Visium-HD format}
\usage{
visiumize(
  data = NULL,
  style = c("spots", "bins"),
  label = NULL,
  bin_res = NULL,
  nb.method = c("Q", "C", "D", "K", "M"),
  k = 20,
  d.max = 40,
  soi = F,
  layers = 1
)
}
\arguments{
\item{data}{a Seurat object containing Kandinsky data slot}

\item{style}{character, define the final data format. Must be one between 'spots' (Visium) and 'bins' (Visium-HD)}

\item{label}{character string specifying the variable name to be used to defne cell annotation groups}

\item{bin_res}{numeric, Visium-HD bin resolution when \code{style} is set to 'bins'.}

\item{nb.method}{spatial neighbour method to apply to the collapsed data. See \code{kandinsky_init} documentation}

\item{k}{numeric, number of nearest neighbours to be set when 'nb.method = K'}

\item{d.max}{numeric, maximum centroid distance threshold to be set when 'nb.method = C | M'}

\item{soi}{boolean, whether or not filter Delaunay network to keep sphere of influence (SOI) graph. Default is FALSE}

\item{layers}{numeric, number of concentric contiguous layers to include in spot neighbourhood. Only Applied when 'nb.method = Q'. Default is 1}
}
\value{
a new Seurat object containing newly generated Visium/Visium-HD data and associated Kandinsky data
}
\description{
This function collapses single-cell gene expression data into spots or bins, depending on
the desired format.
}
