% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Utils.R
\name{get_visium_textures}
\alias{get_visium_textures}
\title{Extract texture features from Visium image}
\usage{
get_visium_textures(visium, aggr.factor = 15, is.hd = F)
}
\arguments{
\item{visium}{a Visium Seurat object containing Kandinsky data (`KanData()`)}

\item{aggr.factor}{numeric argument indicating how many raster cells consider to average RGB intensity and lower raster resolution.}

\item{is.hd}{whether Seurat object contains data from standard Visium (FALSE) or Visium HD (TRUE)
The higher the aggr.factor, the lower the final raster resolution.}
}
\value{
Update Seurat object with a new 'texture' dataframe containing GLCM texture features for each spot stored in the Kandinsky data
}
\description{
Gray-level co-occurrence matrix (GLCM) texture metrics extraction from Visium H&E image
}
\details{
Use the `GLCMTextures` package to extract and summarise GLCM texture features for each tissue spot in the Visium dataset.
Since GLCM texture extraction works with single-layer rasters, the function will extract texture features separately for each
R/G/B raster layer, plus a fourth created by calculating the cell-wise standard deviation across the three RGB layers.
For more information about GLCM texture extraction, see https://github.com/ailich/GLCMTextures
}
