% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Kandinsky_Class_funcs.R
\name{he_mask}
\alias{he_mask}
\alias{he_mask,Seurat-method}
\alias{he_mask,}
\alias{Seurat_he_mask}
\alias{he_mask,SpatRaster-method}
\alias{SpatRaster_he_mask}
\title{Mask Visium H&E image}
\usage{
he_mask(
  object,
  maxres = 2000,
  stretch = F,
  sd_thresh = 9,
  crop = T,
  crop_area = NULL
)

\S4method{he_mask}{Seurat}(
  object,
  maxres = 2000,
  stretch = F,
  sd_thresh = 9,
  crop = T,
  crop_area = NULL
)

\S4method{he_mask}{SpatRaster}(
  object,
  maxres = 2000,
  stretch = F,
  sd_thresh = 9,
  crop = T,
  crop_area = NULL
)
}
\arguments{
\item{object}{a Seurat object containing Kandinsky data (`KanData()`) or an image of class `SpatRaster` created with the `terra` package}

\item{maxres}{numeric value indicating the ideal maximum pixel dimension accepted for Visium H&E image. If image size is bigger than this value, a scale factor will be applied to lower image pixel resolution/size up to the defined maximum value}

\item{stretch}{boolean value, whether or not to apply a linear stretch to increase masked image contrast. Default is FALSE}

\item{sd_thresh}{numeric value indicating the minimum RGB standard deviation accepted to keep an image pixel as part of the final tissue mask. Lower `sd_thresh` values will give more permissive filtering results.}

\item{crop}{boolean value, whether or not to crop the final masked image}

\item{crop_area}{bounding box coordinates to be used to crop masked image. If `crop = T` and `crop_area` is set to `NULL`, Visium spot coordinates will be used to automatically define the bounding box for cropping the image}
}
\value{
Seurat object with new 'masked_img' slot within the Kandinsky data if object is a Seurat object, or a masked `Spatraster` image
}
\description{
create a mask image to remove image background from Visium H&E images
}
\details{
Additional details...
This function will apply a variance-based threshold to the RGB channels of the H&E image stored within the Kandinsky data
to keep only the area of image covering the tissue sample and to remove the rest of the image background. The masking strategy relies on the fact
that RGB values from H&E staining will usually show a higher standard deviation compared to the gray-ish Visium slide background.
}
