% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Hot_cold_spot_analysis_.R
\name{hotspot_overlap}
\alias{hotspot_overlap}
\title{define overlap between hot/cold spots of two variables}
\usage{
hotspot_overlap(seurat = NULL, feat.1 = NULL, feat.2 = NULL, ...)
}
\arguments{
\item{seurat}{a Seurat object containing Kandinsky data (`KanData()`)}

\item{feat.1}{character string specifying the name of the first feature to consider}

\item{feat.2}{character string specifying the name of the second feature to consider}

\item{...}{arguments passed to hotspot_analysis function}
}
\value{
updated Seurat object with a new annotation column reporting the type of overlap between hot/cold spot annotations:
- 'Hot-Hot': a cell/spot is annotated as 'hot' for both the considered variables
- 'Cold-Cold': a cell/spot is annotated as 'cold' for both the considered variables
- 'Cold-Hot': a cell/spot is annotated as 'cold' for one variable and 'hot' for the other
- 'Hot': a cell/spot is annotated as 'hot' only for one of the two variable (the other is annotated as non-significant 'NS')
- 'Cold': a cell/spot is annotated as 'cold' only for one of the two variable (the other is annotated as non-significant 'NS')
- 'NS': a cell/spot is not annotated as either 'hot' or 'cold' for any of the two variables
}
\description{
Given two numeric variables, this function will define hot and cold spots for each variable,
and it will annotate cells/spots according to the presence/absence of overlap between the two hot/cold spot annotations.
If a hot/cold spot annotation for one or both variables already exist,hotspot analysis will not be rerun for that variable(s)
}
