% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Colocalisation.R
\name{jc_coloc}
\alias{jc_coloc}
\title{JC colocalization test and plot using Seurat/Kandinsky object}
\usage{
jc_coloc(
  seurat = NULL,
  label = NULL,
  return.mat = F,
  wc.type = c("mat", "listw")
)
}
\arguments{
\item{seurat}{a Seurat object containing Kandinsky data slot}

\item{label}{name of the metadata column variable containing cell type annotation to be used for join count test}

\item{return.mat}{boolean, whether returning only colocalization plot (FALSE) or also a dataframe containing colocalization test results}

\item{wc.type}{input type to use to extract summary weight constants used to calculate join count statistics. Must be one of the following: "listw", "mat"}
}
\value{
a colocalization plot (when return.mat is FALSE) or a list containing colocalization plot and results in a tabular format
}
\description{
Call multi_jc() function to perform colocalization test, and plot results in the form of a heatmap
}
