% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Kandinsky_Class_funcs.R
\name{KanData}
\alias{KanData}
\alias{KanData,Kandinsky-method}
\alias{KanData,Seurat-method}
\alias{KanData<-,Kandinsky-method}
\alias{KanData<-,Seurat-method}
\title{Show Kandinsky data}
\usage{
KanData(object, which = NULL)

\S4method{KanData}{Kandinsky}(object, which = NULL)

\S4method{KanData}{Seurat}(object, which = NULL)

\S4method{KanData}{Kandinsky}(object, which = NULL) <- value

\S4method{KanData}{Seurat}(object, which = NULL) <- value
}
\arguments{
\item{object}{Seurat or Kandinsky object}

\item{which}{name of Kandinksy slots to extract. If `NULL`, all Kandinsky data will be extracted}

\item{value}{object or variable to assign to the selected Kandinksy slot}
}
\description{
Extract all Kandinsky data or specific slots from Seurat object
}
