% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Prepare_seurat_data.R
\name{prepare_visiumHD_seurat}
\alias{prepare_visiumHD_seurat}
\title{Prepare 10X Visium HD Seurat data}
\usage{
prepare_visiumHD_seurat(
  path = NULL,
  dataset.id = "X",
  img.res = c("full", "high", "low"),
  fullres_path = NULL,
  binsize = c(2, 8, 16)
)
}
\arguments{
\item{path}{character string specifying the path to the Visium input files directory}

\item{dataset.id}{character string that will be used to name the output Seurat object identity}

\item{img.res}{character string specifying which version of H&E Visium image will be loaded into the Seurat data. Must be one of the following: `low`, `high`, `full`, where `full` refers to the original H&E full-resolution tiff image. If set to `full`, lowres image will be loaded into Seurat, but the path to the full res tif image will be also stored in the final object.}

\item{fullres_path}{character string specifying the path to the Visium full resolution H&E tif image. The full image path will be stored into the final Seurat object when `img.res` is set to `full`, and will be used by `Kandinsky` functions for downstream analyses}

\item{binsize}{numeric vector specifying the binsize(s) to be used to build the Seurat object. Can be any combination of `2`, `8`, and `16` sizes}
}
\value{
a Seurat object containing Visium count matrix, metadata, and additional files/file paths to be used for downstream analysis with Kandinsky functions
}
\description{
Initial formatting of Visium data to work with Seurat and Kandinsky
}
\details{
This function will use 10X Visium raw input data to build a new Seurat object. All Visium input files must be stored in the same folder that will be specified to call the function
}
\seealso{
Other prepare_data: 
\code{\link{prepare_cosmx_seurat}()},
\code{\link{prepare_merscope_seurat}()},
\code{\link{prepare_proseg_seurat}()},
\code{\link{prepare_visium_seurat}()},
\code{\link{prepare_xenium_seurat}()}
}
\concept{prepare_data}
